GetUsersIdmResourceTest test is extracted
